InteractivePad/Host/requirements.txt
pyserial  # For listener_serial and other UART-based modules
pyautogui # Send keystrokes
